+++ app/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "庄周吃鱼",
+  "name": "庄周吃鱼",
   "icons": [
     {
       "src": "favicon.ico",
